feat: expose ILayoutProps for better layout customization experience
