+++ docs/docs/convention-routing.zh-CN.md
@@ -130,6 +130,16 @@ Umi 里约定目录下有 `_layout.tsx` 时会生成嵌套路由，以 `_layout.
 ]
 ```
 
+一个自定义的全局 `layout` 如下：
+
+```tsx
+import { ILayoutProps } from 'umi'
+
+export default function Layout({ children, location, route, history, match }: ILayoutProps) {
+  return children
+}
+```
+
 ## 不同的全局 layout
 
 你可能需要针对不同路由输出不同的全局 layout，Umi 不支持这样的配置，但你仍可以在 `src/layouts/index.tsx` 中对 `location.path` 做区分，渲染不同的 layout 。
